refactor: `patchMarkUserVerifiedHandler`
- moved to `user.controller.ts`
