+++ src/controllers/admin.controller.ts
@@ -1,16 +1,8 @@
 import { Request, Response } from "express";
 import { StatusCodes } from "http-status-codes";
-import {
-	DeleteUserSchema,
-	PatchMarkUserVerifiedSchema,
-} from "../schemas/admin.schema";
-import { 
-	deleteUserByIdService,
-	findUsersService,
-	updateUserByIdService,
-} from "../services/user.service";
+import { DeleteUserSchema } from "../schemas/admin.schema";
+import { deleteUserByIdService, findUsersService } from "../services/user.service";
 import logger from "../utils/logger.util";
-
 
 /**
  * This controller will get all users from database
@@ -20,7 +12,7 @@ import logger from "../utils/logger.util";
  *
  * @author aayushchugh, is-itayush
  */
- export const getAllUsersHandler = async (req: Request, res: Response) => {
+export const getAllUsersHandler = async (req: Request, res: Response) => {
 	try {
 		const records = await findUsersService();
 
@@ -62,42 +54,6 @@ export const deleteUserHandler = async (
 
 		return res.status(StatusCodes.OK).json({
 			message: "User deleted successfully",
-		});
-	} catch (err) {
-		logger.error(err);
-
-		return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
-			error: "Internal Server Error",
-		});
-	}
-};
-
-/**
- * This controller will mark user as verified.
- * this can be used by admin to mark any user as verified
- *
- * @param req request
- * @param res response
- *
- * @author aayushchugh, is-it-ayush
- */
-export const patchMarkUserVerifiedHandler = async (
-	req: Request<PatchMarkUserVerifiedSchema["params"]>,
-	res: Response,
-) => {
-	const { id } = req.params;
-
-	try {
-		const verifiedUser = await updateUserByIdService(id, { verified: true });
-
-		if (!verifiedUser) {
-			return res.status(StatusCodes.NOT_FOUND).json({
-				error: "User not found",
-			});
-		}
-
-		return res.status(StatusCodes.OK).json({
-			message: "User verified successfully",
 		});
 	} catch (err) {
 		logger.error(err);
