+++ src/controllers/user.controller.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from "express";
 import { StatusCodes } from "http-status-codes";
-import { PatchMarkUserAdminSchema } from "../schemas/admin.schema";
+import { PatchMarkUserAdminSchema, PatchMarkUserVerifiedSchema } from "../schemas/admin.schema";
 
 import { PatchUserSchema } from "../schemas/user.schema";
 import { findUserByUsernameService, updateUserByIdService } from "../services/user.service";
@@ -42,6 +42,42 @@ export const patchUserHandler = async (
 
 		return res.status(StatusCodes.OK).json({
 			message: "User updated successfully",
+		});
+	} catch (err) {
+		logger.error(err);
+
+		return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({
+			error: "Internal Server Error",
+		});
+	}
+};
+
+/**
+ * This controller will mark user as verified.
+ * this can be used by admin to mark any user as verified
+ *
+ * @param req request
+ * @param res response
+ *
+ * @author aayushchugh, is-it-ayush
+ */
+export const patchMarkUserVerifiedHandler = async (
+	req: Request<PatchMarkUserVerifiedSchema["params"]>,
+	res: Response,
+) => {
+	const { id } = req.params;
+
+	try {
+		const verifiedUser = await updateUserByIdService(id, { verified: true });
+
+		if (!verifiedUser) {
+			return res.status(StatusCodes.NOT_FOUND).json({
+				error: "User not found",
+			});
+		}
+
+		return res.status(StatusCodes.OK).json({
+			message: "User verified successfully",
 		});
 	} catch (err) {
 		logger.error(err);
